Move printDiffOrStringify to jest-matcher-utils
