+++ .eslintignore
@@ -4,7 +4,7 @@ bin/
 flow-typed/**
 packages/*/build/**
 packages/*/build-es5/**
-packages/expect/src/cleanupSemantic.ts
+packages/jest-matcher-utils/src/cleanupSemantic.ts
 website/blog
 website/build
 website/node_modules
