+++ packages/expect/tsconfig.json
@@ -5,7 +5,6 @@
     "outDir": "build"
   },
   "references": [
-    {"path": "../diff-sequences"},
     {"path": "../jest-get-type"},
     {"path": "../jest-matcher-utils"},
     {"path": "../jest-message-util"},
